# public/openapi.yaml
openapi: 3.0.3
info:
  title: Anikoto Scraper API
  description: |
    A high-performance REST API for scraping anime data from **anikoto.net** — built with Next.js, Cheerio, and in-memory caching.

    ## Response Format
    All endpoints return a consistent JSON envelope:
    ```json
    { "ok": true, "data": { ... }, "cached": true }
    ```
    On error:
    ```json
    { "ok": false, "message": "Error description" }
    ```

    ## Caching
    | Endpoint | TTL |
    |----------|-----|
    | `/api/home` | 5 minutes |
    | `/api/anime/:slug` | 30 minutes |
    | `/api/search` | 2 minutes |
    | `/api/filter` | 5 minutes |
    | `/api/schedule` | 1 hour |
    | Episodes | 10 minutes |

    Add `?refresh=1` to any cached endpoint to bypass the cache.
  version: "1.0.0"
  contact:
    name: Teramoto
servers:
  - url: https://anikoto-scrap-api.vercel.app/api
    description: Production server
  - url: /api
    description: Local development server
...
paths:
...
  /anime/{slug}:
    get:
      tags: [Anime]
      summary: Anime detail
      description: Get full anime info including title, synopsis, genres, studios, MAL score, episode count and status.
      operationId: getAnimeDetail
      parameters:
        - $ref: '#/components/parameters/slug'
        - $ref: '#/components/parameters/refresh'
      responses:
        '200':
          description: Anime detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  cached:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/AnimeDetail'
              example:
                ok: true
                cached: true
                data:
                  id: one-piece-odmau
                  slug: one-piece-odmau
                  title: One Piece
                  type: TV
                  status: Currently Airing
                  malScore: 8.9
                  genres: [Action, Adventure, Comedy]
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Error'
...
  /anime/{slug}/related:
    get:
      tags: [Anime]
      summary: Related anime listing
      description: Returns related anime (watch order/sequels/prequels/etc.) for a given anime slug.
      operationId: getAnimeRelated
      parameters:
        - $ref: '#/components/parameters/slug'
        - $ref: '#/components/parameters/refresh'
      responses:
        '200':
          description: Related anime list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  cached:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RelatedAnime'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Error'
  /anime/{slug}/recommendations:
    get:
      tags: [Anime]
      summary: Anime recommendations listing
      description: Returns recommended anime cards for a given anime slug.
      operationId: getAnimeRecommendations
      parameters:
        - $ref: '#/components/parameters/slug'
        - $ref: '#/components/parameters/refresh'
      responses:
        '200':
          description: Recommendations listing
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  cached:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnimeCard'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Error'
...
components:
  parameters:
    slug:
      name: slug
      in: path
      required: true
      schema:
        type: string
      description: Anime slug from the URL (e.g. one-piece-odmau)
      example: one-piece-odmau
    refresh:
      name: refresh
      in: query
      required: false
      schema:
        type: string
        enum: ["1"]
      description: Set to `1` to bypass cache and force a fresh scrape
  schemas:
...
    EpisodeStatus:
      type: object
      properties:
        sub:
          type: integer
          nullable: true
        dub:
          type: integer
          nullable: true
        total:
          type: integer
          nullable: true
    AnimeCard:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        titleJp:
          type: string
        image:
          type: string
          format: uri
        href:
          type: string
        type:
          type: string
        episodes:
          $ref: '#/components/schemas/EpisodeStatus'
        date:
          type: string
        score:
          type: number
        totalEpisodes:
          type: integer

    # ─── Home ──────────────────────────────────────────────────────
...
    RelatedAnime:
      type: object
      properties:
        id:
          type: string
        malId:
          type: string
        title:
          type: string
        titleJp:
          type: string
        image:
          type: string
          format: uri
        relation:
          type: string
        href:
          type: string
        slug:
          type: string
    Episode:
      type: object
      properties:
        number:
          type: string
        title:
          type: string
        href:
          type: string
        id:
          type: string
        dataIds:
          type: string
        hasDub:
          type: boolean
        hasSub:
          type: boolean
    AnimeEpisodes:
      type: object
      properties:
        animeId:
          type: string
        slug:
          type: string
        episodes:
          type: array
          items:
            $ref: '#/components/schemas/Episode'
    AnimeDetail:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        titleJp:
          type: string
        alternativeTitles:
          type: array
          items:
            type: string
        image:
          type: string
          format: uri
        rating:
          type: string
        quality:
          type: string
        hasDub:
          type: boolean
        hasSub:
          type: boolean
        synopsis:
          type: string
        type:
          type: string
        premiered:
          type: string
        aired:
          type: string
        status:
          type: string
        genres:
          type: array
          items:
            type: string
        malScore:
          type: number
        duration:
          type: string
        episodeCount:
          type: integer
        studios:
          type: array
          items:
            type: string
        producers:
          type: array
          items:
            type: string
        watchUrl:
          type: string
        episodes:
          $ref: '#/components/schemas/AnimeEpisodes'

    # ─── Search / Filter ───────────────────────────────────────────
...
    ApiError:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
          example: false
        message:
          type: string
          example: "Internal server error"
  responses:
    Error:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            ok: false
            message: Failed to fetch data from source
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            ok: false
            message: Anime not found